Move formatters to their own package
