+++ scripts/documentset/formatter/index.js
@@ -1,0 +1,7 @@
+// Export document formatters. Each formatter transforms its arguments (lines, speakers, mentions, userTz) into
+// an object with the spec {body, speakers, mentions}.
+
+module.exports = {
+  quote: require('./quote'),
+  lim: require('./lim')
+}
