+++ scripts/quote.js
@@ -3,32 +3,7 @@
 
 const {createDocumentSet} = require('./documentset')
 const {Admin, QuotePontiff, PoetLaureate} = require('./roles')
-
-const limFormatter = (lines, speakers, mentions) => {
-  let body = lines.map(line => `> ${line.text}`).join('\n')
-
-  const atSpeakers = Array.from(speakers, speaker => '@' + speaker)
-
-  switch (atSpeakers.length) {
-    case 0:
-      body += '\n\n  - _anonymous_'
-      break
-    case 1:
-      body += `\n\n  - by @${atSpeakers[0]}`
-      break
-    case 2:
-      body += `\n\n  - a collaboration by ${atSpeakers.join(' and ')}`
-      break
-    default:
-      const allButLast = atSpeakers.slice(0, atSpeakers.length - 1)
-      const last = atSpeakers[atSpeakers.length - 1]
-
-      body += `\n\n - by ${allButLast.join(', ')} and ${last}`
-      break
-  }
-
-  return {body, speakers, mentions: []}
-}
+const formatter = require('./documentset/formatter')
 
 module.exports = function (robot) {
   if (!robot.postgres) {
@@ -51,7 +26,7 @@ module.exports = function (robot) {
   createDocumentSet(robot, 'lim', {
     add: {
       role: PoetLaureate,
-      formatter: limFormatter
+      formatter: formatter.lim
     },
     query: true,
     count: true,
